refactor: remove all docstrings from schemas and models
The excessive use of docstrings proved detrimental to the project's goal of providing a boilerplate for rapid API development in FastAPI, adhering to the principle of 'A backend for perfectionists with deadlines'. Only the docstrings of base models in 'backend/src/core/common/models.py' have been retained.
Adjust migration script to reorder database columns in a more logical manner.